requirements.txt
langchain
langchain-community
langchain-openai
langgraph
langchain-qdrant
qdrant-client
streamlit
pypdf
requests
python-dotenv
langsmith
pytest
openai
